Prevent incorrect connectInject config
Connect injection requires clients to be enabled and the gRPC port to be
enabled. These requirements are documented on the website although not
in the values.yaml itself. If users do not enable clients, their init
pods will fail. If users do not enable gRPC, their sidecars will be
reported as unhealthy. Both these outcomes are hard to debug.

This change makes it harder to get into this misconfiguration. It will
cause a helm error if clients are not enabled when connect is enabled
and it will automatically enable gRPC if connect is enabled.

This is a breaking change, however any users that had connect enabled
but no clients or gRPC not enabled would be running into errors so it's
not going to break anything for anyone that isn't already in a broken
state.